format

+++ example/graph_executor_with_tvm/tests/test_tvm_op.py
@@ -2,8 +2,8 @@ import numpy as np
 import tvm
 from dlsys import autodiff, tvm_op
 
-tgt_host="llvm"
-tgt="llvm"
+tgt_host = "llvm"
+tgt = "llvm"
 dtype = "float32"
 ctx = tvm.context(tgt, 0)
 
@@ -29,7 +29,8 @@ def test_matrix_elementwise_add_by_const():
     y = np.zeros(shape).astype(dtype)
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
-    elemwise_add_by_const = tvm_op.make_elemwise_add_by_const(shape, const_val, tgt, tgt_host, "elem_add_by_const")
+    elemwise_add_by_const = tvm_op.make_elemwise_add_by_const(
+        shape, const_val, tgt, tgt_host, "elem_add_by_const")
     elemwise_add_by_const(arr_x, arr_y)
     y = arr_y.asnumpy()
     np.testing.assert_allclose(x + const_val, y, rtol=1e-5)
@@ -56,7 +57,8 @@ def test_matrix_elementwise_mul_by_const():
     y = np.zeros(shape).astype(dtype)
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
-    elemwise_mul_by_const = tvm_op.make_elemwise_mul_by_const(shape, const_val, tgt, tgt_host, "elem_mul_by_const")
+    elemwise_mul_by_const = tvm_op.make_elemwise_mul_by_const(
+        shape, const_val, tgt, tgt_host, "elem_mul_by_const")
     elemwise_mul_by_const(arr_x, arr_y)
     y = arr_y.asnumpy()
     np.testing.assert_allclose(x * const_val, y, rtol=1e-5)
@@ -72,8 +74,9 @@ def test_matrix_multiply():
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
     arr_z = tvm.nd.array(z, ctx=ctx)
-   
-    matrix_mul = tvm_op.make_matrix_mul(shapeX, False, shapeY, False, tgt, tgt_host, "matrix_mul")
+
+    matrix_mul = tvm_op.make_matrix_mul(
+        shapeX, False, shapeY, False, tgt, tgt_host, "matrix_mul")
     matrix_mul(arr_x, arr_y, arr_z)
     z = arr_z.asnumpy()
     np.testing.assert_allclose(np.dot(x, y), z, rtol=1e-5)
@@ -88,14 +91,15 @@ def test_matrix_multiply():
     arr_y = tvm.nd.array(y, ctx=ctx)
     arr_z = tvm.nd.array(z, ctx=ctx)
 
-    matrix_mul = tvm_op.make_matrix_mul(shapeX, False, shapeY, True, tgt, tgt_host, "matrix_mul")
+    matrix_mul = tvm_op.make_matrix_mul(
+        shapeX, False, shapeY, True, tgt, tgt_host, "matrix_mul")
     matrix_mul(arr_x, arr_y, arr_z)
     z = arr_z.asnumpy()
     np.testing.assert_allclose(np.dot(x, np.transpose(y)), z, rtol=1e-5)
-    
+
     shapeX = (500, 1000)
     shapeY = (500, 2000)
-    shapeZ = (1000, 2000)   
+    shapeZ = (1000, 2000)
     x = np.random.uniform(0, 10, size=shapeX).astype(dtype)
     y = np.random.uniform(0, 10, size=shapeY).astype(dtype)
     z = np.zeros(shapeZ).astype(dtype)
@@ -103,14 +107,15 @@ def test_matrix_multiply():
     arr_y = tvm.nd.array(y, ctx=ctx)
     arr_z = tvm.nd.array(z, ctx=ctx)
 
-    matrix_mul = tvm_op.make_matrix_mul(shapeX, True, shapeY, False, tgt, tgt_host, "matrix_mul")
+    matrix_mul = tvm_op.make_matrix_mul(
+        shapeX, True, shapeY, False, tgt, tgt_host, "matrix_mul")
     matrix_mul(arr_x, arr_y, arr_z)
     z = arr_z.asnumpy()
     np.testing.assert_allclose(np.dot(np.transpose(x), y), z, rtol=1e-5)
-    
+
     shapeX = (500, 1000)
     shapeY = (2000, 500)
-    shapeZ = (1000, 2000)   
+    shapeZ = (1000, 2000)
     x = np.random.uniform(0, 10, size=shapeX).astype(dtype)
     y = np.random.uniform(0, 10, size=shapeY).astype(dtype)
     z = np.zeros(shapeZ).astype(dtype)
@@ -118,42 +123,45 @@ def test_matrix_multiply():
     arr_y = tvm.nd.array(y, ctx=ctx)
     arr_z = tvm.nd.array(z, ctx=ctx)
 
-    matrix_mul = tvm_op.make_matrix_mul(shapeX, True, shapeY, True, tgt, tgt_host, "matrix_mul")
+    matrix_mul = tvm_op.make_matrix_mul(
+        shapeX, True, shapeY, True, tgt, tgt_host, "matrix_mul")
     matrix_mul(arr_x, arr_y, arr_z)
     z = arr_z.asnumpy()
-    np.testing.assert_allclose(np.dot(np.transpose(x), np.transpose(y)), z, rtol=1e-5)
+    np.testing.assert_allclose(
+        np.dot(np.transpose(x), np.transpose(y)), z, rtol=1e-5)
 
 
 def test_conv2d():
     # im2col and np_conv2d are helper functions
     def im2col(X, filter_H, filter_W, padding, stride):
-      N, C, H, W = X.shape
-      assert (H + 2 * padding - filter_H) % stride == 0
-      assert (W + 2 * padding - filter_W) % stride == 0
-      out_H = (H + 2 * padding - filter_H) / stride + 1
-      out_W = (W + 2 * padding - filter_W) / stride + 1
+        N, C, H, W = X.shape
+        assert (H + 2 * padding - filter_H) % stride == 0
+        assert (W + 2 * padding - filter_W) % stride == 0
+        out_H = (H + 2 * padding - filter_H) / stride + 1
+        out_W = (W + 2 * padding - filter_W) / stride + 1
 
-      y_row_size = C * filter_H * filter_W
-      y_col_size = out_H * out_W
-      y_shape = (N, y_row_size, y_col_size)
-      Y = np.empty(y_shape, dtype = X.dtype)
+        y_row_size = C * filter_H * filter_W
+        y_col_size = out_H * out_W
+        y_shape = (N, y_row_size, y_col_size)
+        Y = np.empty(y_shape, dtype=X.dtype)
 
-      for batch_index in range(N):
-        for col_index in range(y_col_size):
-          out_y = col_index / out_W
-          out_x = col_index % out_W
-          in_y = out_y * stride - padding
-          in_x = out_x * stride - padding
-          row_idx = 0
-          for c in range(0, C):
-            for y in range(in_y, in_y + filter_H):
-              for x in range(in_x, in_x + filter_W):
-                if (x < 0 or x >= W or y < 0 or y >= H):
-                  Y[batch_index, row_idx, col_index] = 0
-                else:
-                  Y[batch_index, row_idx, col_index] = X[batch_index, c, y, x]
-                row_idx += 1
-      return Y
+        for batch_index in range(N):
+            for col_index in range(y_col_size):
+                out_y = col_index / out_W
+                out_x = col_index % out_W
+                in_y = out_y * stride - padding
+                in_x = out_x * stride - padding
+                row_idx = 0
+                for c in range(0, C):
+                    for y in range(in_y, in_y + filter_H):
+                        for x in range(in_x, in_x + filter_W):
+                            if (x < 0 or x >= W or y < 0 or y >= H):
+                                Y[batch_index, row_idx, col_index] = 0
+                            else:
+                                Y[batch_index, row_idx,
+                                    col_index] = X[batch_index, c, y, x]
+                            row_idx += 1
+        return Y
 
     def np_conv2d(X, Filter, padding=0, stride=1):
         """Implement a conv2d as a matrix multiply after im2col."""
@@ -177,10 +185,10 @@ def test_conv2d():
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_f = tvm.nd.array(f, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
-   
+
     conv2d = tvm_op.make_conv2d(shapeX, shapeF, tgt, tgt_host, "conv2d")
     conv2d(arr_x, arr_f, arr_y)
-    y = arr_y.asnumpy()   
+    y = arr_y.asnumpy()
     np.testing.assert_allclose(np_conv2d(x, f), y, rtol=1e-5)
 
 
@@ -204,7 +212,8 @@ def test_relu_gradient():
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_grad_x = tvm.nd.array(grad_x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
-    relu_gradient = tvm_op.make_relu_gradient(shape, tgt, tgt_host, "relu_gradient")
+    relu_gradient = tvm_op.make_relu_gradient(
+        shape, tgt, tgt_host, "relu_gradient")
     relu_gradient(arr_x, arr_grad_x, arr_y)
     y = arr_y.asnumpy()
     np.testing.assert_allclose(((x > 0) * grad_x).astype(dtype), y)
@@ -216,7 +225,8 @@ def test_softmax():
     y = np.zeros(shape).astype(dtype)
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
-    matrix_softmax = tvm_op.make_matrix_softmax(shape, tgt, tgt_host, "matrix_softmax")
+    matrix_softmax = tvm_op.make_matrix_softmax(
+        shape, tgt, tgt_host, "matrix_softmax")
     matrix_softmax(arr_x, arr_y)
     y = arr_y.asnumpy()
     np.testing.assert_allclose(autodiff.softmax_func(x), y, rtol=1e-5)
@@ -230,7 +240,8 @@ def test_softmax_cross_entropy():
     arr_y = tvm.nd.array(y, ctx=ctx)
     arr_y_ = tvm.nd.array(y_, ctx=ctx)
     arr_out = tvm.nd.array(out, ctx=ctx)
-    matrix_softmax_cross_entropy = tvm_op.make_matrix_softmax_cross_entropy(shape, tgt, tgt_host, "softmax_cross_entropy")
+    matrix_softmax_cross_entropy = tvm_op.make_matrix_softmax_cross_entropy(
+        shape, tgt, tgt_host, "softmax_cross_entropy")
     matrix_softmax_cross_entropy(arr_y, arr_y_, arr_out)
     out = arr_out.asnumpy()
     # numpy calculation
@@ -247,9 +258,10 @@ def test_reduce_sum_axis_zero():
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
 
-    reduce_sum_axis_zero = tvm_op.make_reduce_sum_axis_zero(shape, tgt, tgt_host, "reduce_sum_axis_zero")
+    reduce_sum_axis_zero = tvm_op.make_reduce_sum_axis_zero(
+        shape, tgt, tgt_host, "reduce_sum_axis_zero")
     reduce_sum_axis_zero(arr_x, arr_y)
-    
+
     y = arr_y.asnumpy()
     np.testing.assert_allclose(np.sum(x, axis=0), y, rtol=1e-5)
 
@@ -261,8 +273,8 @@ def test_broadcast_to():
     y = np.zeros(to_shape).astype(dtype)
     arr_x = tvm.nd.array(x, ctx=ctx)
     arr_y = tvm.nd.array(y, ctx=ctx)
-    broadcast_to = tvm_op.make_broadcast_to(shape, to_shape, tgt, tgt_host, "broadcast_to")
+    broadcast_to = tvm_op.make_broadcast_to(
+        shape, to_shape, tgt, tgt_host, "broadcast_to")
     broadcast_to(arr_x, arr_y)
     y = arr_y.asnumpy()
     np.testing.assert_allclose(np.broadcast_to(x, to_shape), y)
-
